# build/public/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"ApiResponse": {
				"properties": {
					"status": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {},
					"code": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"status",
					"message",
					"code"
				],
				"type": "object",
				"additionalProperties": false
			},
			"InitMultipartRequest": {
				"properties": {
					"fileName": {
						"type": "string"
					},
					"fileType": {
						"type": "string"
					}
				},
				"required": [
					"fileName",
					"fileType"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SignPartRequest": {
				"properties": {
					"key": {
						"type": "string"
					},
					"uploadId": {
						"type": "string"
					},
					"partNumber": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"key",
					"uploadId",
					"partNumber"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CompleteMultipartRequest": {
				"properties": {
					"key": {
						"type": "string"
					},
					"uploadId": {
						"type": "string"
					},
					"parts": {
						"items": {
							"properties": {
								"PartNumber": {
									"type": "number",
									"format": "double"
								},
								"ETag": {
									"type": "string"
								}
							},
							"required": [
								"PartNumber",
								"ETag"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"key",
					"uploadId",
					"parts"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"Bearer": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header",
				"description": "JWT Authorization header using the Bearer scheme."
			}
		}
	},
	"info": {
		"title": "heal_api_3001",
		"version": "1.0.0",
		"description": "Boilerplate apis",
		"license": {
			"name": "license"
		},
		"contact": {
			"name": "STS"
		}
	},
	"paths": {
...
		"/user/auth/details": {
			"get": {
				"operationId": "GetUserDetails",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Get User info",
				"tags": [
					"User Auth Routes"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": []
			}
		},
...
		"/admin/user/list": {
			"get": {
				"operationId": "GetUsersList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Get User List",
				"tags": [
					"Admin User Routes"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "sort_column",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sort_direction",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "search_key",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/admin/user/details": {
			"get": {
				"operationId": "GetUserDetails",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Get User Details",
				"tags": [
					"Admin User Routes"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "user_id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/admin/user/status": {
			"put": {
				"operationId": "UpdateUserStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Update User Status\n1 for active 2 for delete 3 for deactivate",
				"tags": [
					"Admin User Routes"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"status": {
										"type": "number",
										"format": "double"
									},
									"user_id": {
										"type": "string"
									}
								},
								"required": [
									"status",
									"user_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/admin/user/dashboard": {
			"get": {
				"operationId": "GetDashboardData",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Get Dashboard data",
				"tags": [
					"Admin User Routes"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "past_day",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/admin/user/upload/multipart/init": {
			"post": {
				"operationId": "InitiateMultipartUpload",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "1. Initiate Multipart Upload",
				"tags": [
					"Admin User Routes"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/InitMultipartRequest"
							}
						}
					}
				}
			}
		},
		"/admin/user/upload/multipart/sign-part": {
			"post": {
				"operationId": "SignMultipartPart",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "2. Sign Part",
				"tags": [
					"Admin User Routes"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SignPartRequest"
							}
						}
					}
				}
			}
		},
		"/admin/user/upload/multipart/complete": {
			"post": {
				"operationId": "CompleteMultipartUpload",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "3. Complete Multipart",
				"tags": [
					"Admin User Routes"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CompleteMultipartRequest"
							}
						}
					}
				}
			}
		},
		"/admin/user/upload/process_file_admin": {
			"post": {
				"operationId": "ProcessFileAdmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Process File Admin",
				"tags": [
					"Admin User Routes"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"duration": {
										"type": "number",
										"format": "double"
									},
									"s3_key": {
										"type": "string"
									}
								},
								"required": [
									"s3_key"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}